Struct aws_sdk_mediapackage::client::fluent_builders::CreateOriginEndpoint

pub struct CreateOriginEndpoint<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> CreateOriginEndpoint<C>

pub async fn send(
self
) -> Result<CreateOriginEndpointOutput, SdkError<CreateOriginEndpointError>> where
C: SmithyConnector,

pub fn authorization(self, input: Authorization) -> Self

CDN Authorization credentials

pub fn set_authorization(self, input: Option<Authorization>) -> Self

pub fn channel_id(self, input: impl Into<String>) -> Self

The ID of the Channel that the OriginEndpoint will be associated with. This cannot be changed after the OriginEndpoint is created.

pub fn set_channel_id(self, input: Option<String>) -> Self

pub fn cmaf_package(self, input: CmafPackageCreateOrUpdateParameters) -> Self

A Common Media Application Format (CMAF) packaging configuration.

pub fn set_cmaf_package(
self,
input: Option<CmafPackageCreateOrUpdateParameters>
) -> Self

pub fn dash_package(self, input: DashPackage) -> Self

A Dynamic Adaptive Streaming over HTTP (DASH) packaging configuration.

pub fn set_dash_package(self, input: Option<DashPackage>) -> Self

pub fn description(self, input: impl Into<String>) -> Self

A short text description of the OriginEndpoint.

pub fn set_description(self, input: Option<String>) -> Self

pub fn hls_package(self, input: HlsPackage) -> Self

An HTTP Live Streaming (HLS) packaging configuration.

pub fn set_hls_package(self, input: Option<HlsPackage>) -> Self

pub fn id(self, input: impl Into<String>) -> Self

The ID of the OriginEndpoint. The ID must be unique within the region and it cannot be changed after the OriginEndpoint is created.

pub fn set_id(self, input: Option<String>) -> Self

pub fn manifest_name(self, input: impl Into<String>) -> Self

A short string that will be used as the filename of the OriginEndpoint URL (defaults to “index”).

pub fn set_manifest_name(self, input: Option<String>) -> Self

pub fn mss_package(self, input: MssPackage) -> Self

A Microsoft Smooth Streaming (MSS) packaging configuration.

pub fn set_mss_package(self, input: Option<MssPackage>) -> Self

pub fn origination(self, input: Origination) -> Self

Control whether origination of video is allowed for this OriginEndpoint. If set to ALLOW, the OriginEndpoint may by requested, pursuant to any other form of access control. If set to DENY, the OriginEndpoint may not be requested. This can be helpful for Live to VOD harvesting, or for temporarily disabling origination

pub fn set_origination(self, input: Option<Origination>) -> Self

pub fn startover_window_seconds(self, input: i32) -> Self

Maximum duration (seconds) of content to retain for startover playback. If not specified, startover playback will be disabled for the OriginEndpoint.

pub fn set_startover_window_seconds(self, input: Option<i32>) -> Self

pub fn tags(self, k: impl Into<String>, v: impl Into<String>) -> Self

A collection of tags associated with a resource

pub fn set_tags(self, input: Option<HashMap<String, String>>) -> Self

pub fn time_delay_seconds(self, input: i32) -> Self

Amount of delay (seconds) to enforce on the playback of live content. If not specified, there will be no time delay in effect for the OriginEndpoint.

pub fn set_time_delay_seconds(self, input: Option<i32>) -> Self

pub fn whitelist(self, inp: impl Into<String>) -> Self

A list of source IP CIDR blocks that will be allowed to access the OriginEndpoint.

pub fn set_whitelist(self, input: Option<Vec<String>>) -> Self