Module aws_sdk_mediapackage::error

Modules

forbidden_exception

See ForbiddenException

internal_server_error_exception

See InternalServerErrorException

not_found_exception

See NotFoundException

service_unavailable_exception

See ServiceUnavailableException

too_many_requests_exception

See TooManyRequestsException

unprocessable_entity_exception

See UnprocessableEntityException

Structs

ConfigureLogsError

CreateChannelError

CreateHarvestJobError

CreateOriginEndpointError

DeleteChannelError

DeleteOriginEndpointError

DescribeChannelError

DescribeHarvestJobError

DescribeOriginEndpointError

ForbiddenException

The client is not authorized to access the requested resource.

InternalServerErrorException

An unexpected error occurred.

ListChannelsError

ListHarvestJobsError

ListOriginEndpointsError

ListTagsForResourceError

NotFoundException

The requested resource does not exist.

RotateChannelCredentialsError

RotateIngestEndpointCredentialsError

ServiceUnavailableException

An unexpected error occurred.

TagResourceError

TooManyRequestsException

The client has exceeded their resource or throttling limits.

UnprocessableEntityException

The parameters sent in the request are not valid.

UntagResourceError

UpdateChannelError

UpdateOriginEndpointError

Enums

ConfigureLogsErrorKind

CreateChannelErrorKind

CreateHarvestJobErrorKind

CreateOriginEndpointErrorKind

DeleteChannelErrorKind

DeleteOriginEndpointErrorKind

DescribeChannelErrorKind

DescribeHarvestJobErrorKind

DescribeOriginEndpointErrorKind

ListChannelsErrorKind

ListHarvestJobsErrorKind

ListOriginEndpointsErrorKind

ListTagsForResourceErrorKind

RotateChannelCredentialsErrorKind

RotateIngestEndpointCredentialsErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateChannelErrorKind

UpdateOriginEndpointErrorKind