Module aws_sdk_mediapackage::model

Modules

authorization

See Authorization

channel

See Channel

cmaf_encryption

See CmafEncryption

cmaf_package

See CmafPackage

cmaf_package_create_or_update_parameters

See CmafPackageCreateOrUpdateParameters

dash_encryption

See DashEncryption

dash_package

See DashPackage

egress_access_logs

See EgressAccessLogs

encryption_contract_configuration

See EncryptionContractConfiguration

harvest_job

See HarvestJob

hls_encryption

See HlsEncryption

hls_ingest

See HlsIngest

hls_manifest

See HlsManifest

hls_manifest_create_or_update_parameters

See HlsManifestCreateOrUpdateParameters

hls_package

See HlsPackage

ingest_endpoint

See IngestEndpoint

ingress_access_logs

See IngressAccessLogs

mss_encryption

See MssEncryption

mss_package

See MssPackage

origin_endpoint

See OriginEndpoint

s3_destination

See S3Destination

speke_key_provider

See SpekeKeyProvider

stream_selection

See StreamSelection

Structs

Authorization

CDN Authorization credentials

Channel

A Channel resource configuration.

CmafEncryption

A Common Media Application Format (CMAF) encryption configuration.

CmafPackage

A Common Media Application Format (CMAF) packaging configuration.

CmafPackageCreateOrUpdateParameters

A Common Media Application Format (CMAF) packaging configuration.

DashEncryption

A Dynamic Adaptive Streaming over HTTP (DASH) encryption configuration.

DashPackage

A Dynamic Adaptive Streaming over HTTP (DASH) packaging configuration.

EgressAccessLogs

Configure egress access logging.

EncryptionContractConfiguration

Use encryptionContractConfiguration to configure one or more content encryption keys for your endpoints that use SPEKE 2.0. The encryption contract defines which content keys are used to encrypt the audio and video tracks in your stream. To configure the encryption contract, specify which audio and video encryption presets to use. Note the following considerations when using encryptionContractConfiguration: encryptionContractConfiguration can be used for DASH endpoints that use SPEKE 2.0. SPEKE 2.0 relies on the CPIX 2.3 specification. You must disable key rotation for this endpoint by setting keyRotationIntervalSeconds to 0.

HarvestJob

A HarvestJob resource configuration

HlsEncryption

An HTTP Live Streaming (HLS) encryption configuration.

HlsIngest

An HTTP Live Streaming (HLS) ingest resource configuration.

HlsManifest

A HTTP Live Streaming (HLS) manifest configuration.

HlsManifestCreateOrUpdateParameters

A HTTP Live Streaming (HLS) manifest configuration.

HlsPackage

An HTTP Live Streaming (HLS) packaging configuration.

IngestEndpoint

An endpoint for ingesting source content for a Channel.

IngressAccessLogs

Configure ingress access logging.

MssEncryption

A Microsoft Smooth Streaming (MSS) encryption configuration.

MssPackage

A Microsoft Smooth Streaming (MSS) packaging configuration.

OriginEndpoint

An OriginEndpoint resource configuration.

S3Destination

Configuration parameters for where in an S3 bucket to place the harvested content

SpekeKeyProvider

A configuration for accessing an external Secure Packager and Encoder Key Exchange (SPEKE) service that will provide encryption keys.

StreamSelection

A StreamSelection configuration.

Enums

AdMarkers

AdTriggersElement

AdsOnDeliveryRestrictions

This setting allows the delivery restriction flags on SCTE-35 segmentation descriptors to determine whether a message signals an ad. Choosing “NONE” means no SCTE-35 messages become ads. Choosing “RESTRICTED” means SCTE-35 messages of the types specified in AdTriggers that contain delivery restrictions will be treated as ads. Choosing “UNRESTRICTED” means SCTE-35 messages of the types specified in AdTriggers that do not contain delivery restrictions will be treated as ads. Choosing “BOTH” means all SCTE-35 messages of the types specified in AdTriggers will be treated as ads. Note that Splice Insert messages do not have these flags and are always treated as ads if specified in AdTriggers.

EncryptionMethod

ManifestLayout

Origination

PeriodTriggersElement

PlaylistType

PresetSpeke20Audio

PresetSpeke20Video

Profile

SegmentTemplateFormat

Status

StreamOrder

UtcTiming