Struct aws_sdk_mediapackage::model::dash_package::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for DashPackage

Implementations

impl Builder

pub fn ad_triggers(self, input: impl Into<AdTriggersElement>) -> Self

pub fn set_ad_triggers(self, input: Option<Vec<AdTriggersElement>>) -> Self

pub fn ads_on_delivery_restrictions(
self,
input: AdsOnDeliveryRestrictions
) -> Self

This setting allows the delivery restriction flags on SCTE-35 segmentation descriptors to determine whether a message signals an ad. Choosing “NONE” means no SCTE-35 messages become ads. Choosing “RESTRICTED” means SCTE-35 messages of the types specified in AdTriggers that contain delivery restrictions will be treated as ads. Choosing “UNRESTRICTED” means SCTE-35 messages of the types specified in AdTriggers that do not contain delivery restrictions will be treated as ads. Choosing “BOTH” means all SCTE-35 messages of the types specified in AdTriggers will be treated as ads. Note that Splice Insert messages do not have these flags and are always treated as ads if specified in AdTriggers.

pub fn set_ads_on_delivery_restrictions(
self,
input: Option<AdsOnDeliveryRestrictions>
) -> Self

pub fn encryption(self, input: DashEncryption) -> Self

A Dynamic Adaptive Streaming over HTTP (DASH) encryption configuration.

pub fn set_encryption(self, input: Option<DashEncryption>) -> Self

pub fn manifest_layout(self, input: ManifestLayout) -> Self

Determines the position of some tags in the Media Presentation Description (MPD). When set to FULL, elements like SegmentTemplate and ContentProtection are included in each Representation. When set to COMPACT, duplicate elements are combined and presented at the AdaptationSet level.

pub fn set_manifest_layout(self, input: Option<ManifestLayout>) -> Self

pub fn manifest_window_seconds(self, input: i32) -> Self

Time window (in seconds) contained in each manifest.

pub fn set_manifest_window_seconds(self, input: Option<i32>) -> Self

pub fn min_buffer_time_seconds(self, input: i32) -> Self

Minimum duration (in seconds) that a player will buffer media before starting the presentation.

pub fn set_min_buffer_time_seconds(self, input: Option<i32>) -> Self

pub fn min_update_period_seconds(self, input: i32) -> Self

Minimum duration (in seconds) between potential changes to the Dynamic Adaptive Streaming over HTTP (DASH) Media Presentation Description (MPD).

pub fn set_min_update_period_seconds(self, input: Option<i32>) -> Self

pub fn period_triggers(self, input: impl Into<PeriodTriggersElement>) -> Self

pub fn set_period_triggers(
self,
input: Option<Vec<PeriodTriggersElement>>
) -> Self

pub fn profile(self, input: Profile) -> Self

The Dynamic Adaptive Streaming over HTTP (DASH) profile type. When set to “HBBTV_1_5”, HbbTV 1.5 compliant output is enabled.

pub fn set_profile(self, input: Option<Profile>) -> Self

pub fn segment_duration_seconds(self, input: i32) -> Self

Duration (in seconds) of each segment. Actual segments will be rounded to the nearest multiple of the source segment duration.

pub fn set_segment_duration_seconds(self, input: Option<i32>) -> Self

pub fn segment_template_format(self, input: SegmentTemplateFormat) -> Self

Determines the type of SegmentTemplate included in the Media Presentation Description (MPD). When set to NUMBER_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Number$ media URLs. When set to TIME_WITH_TIMELINE, a full timeline is presented in each SegmentTemplate, with $Time$ media URLs. When set to NUMBER_WITH_DURATION, only a duration is included in each SegmentTemplate, with $Number$ media URLs.

pub fn set_segment_template_format(
self,
input: Option<SegmentTemplateFormat>
) -> Self

pub fn stream_selection(self, input: StreamSelection) -> Self

A StreamSelection configuration.

pub fn set_stream_selection(self, input: Option<StreamSelection>) -> Self

pub fn suggested_presentation_delay_seconds(self, input: i32) -> Self

Duration (in seconds) to delay live content before presentation.

pub fn set_suggested_presentation_delay_seconds(
self,
input: Option<i32>
) -> Self

pub fn utc_timing(self, input: UtcTiming) -> Self

Determines the type of UTCTiming included in the Media Presentation Description (MPD)

pub fn set_utc_timing(self, input: Option<UtcTiming>) -> Self

pub fn utc_timing_uri(self, input: impl Into<String>) -> Self

Specifies the value attribute of the UTCTiming field when utcTiming is set to HTTP-ISO or HTTP-HEAD

pub fn set_utc_timing_uri(self, input: Option<String>) -> Self

pub fn build(self) -> DashPackage

Consumes the builder and constructs a DashPackage